Refactor menu system completely
Removed manual menu creation in favor of a more modular menu system. Added a new Menu interface and related classes to facilitate this change. Made changes to the menu.yml file and the MenuUtil class according to the new system. This will allow menu items to be created more easily and systematically.