Implement trusted user role and permissions system
This update introduces a new user role ("Trusted") in the islandsmp's entities and utils packages. In addition, the permissions of each role have been defined and appropriate methods added to manage these permissions. This includes adding, checking, removing permissions, as well as determining the rank of a user. The display of information related to trusted members and their actions was implemented in the menu.