Add upgrade functionality and respective messages
The commit introduces the functionality for upgrading tiers and resetting the amount of spins in the Island SMP game. It also includes the addition of new messages for when the player upgrades their tier. The GUI has been updated to handle these new features as well. The changes foster game progression and give players more avenues to interact with the game environment.